+++ docs/__init__.py
@@ -1,18 +1,25 @@
 #! /usr/bin/env python
 # -*- coding: utf-8 -*-
+#
+# Interpreter version: python 2.7, 3.4
+#
+
 
 def allSame(s):
-    return not filter(lambda x: x != s[0], s)
+    return not any(filter(lambda x: x != s[0], s))
 
 
 def hasDigit(s):
-    return any(map(lambda x: x.isdigit(), s))
+    return any(char.isdigit() for char in s)
 
 
 def getVersion(data):
+    """
+    Parse version from changelog written in RST format.
+    """
     data = data.splitlines()
-    return filter(
-        lambda x, y:
-            len(x) == len(y) and allSame(y) and hasDigit(x) and "." in x,
-        zip(data, data[1:])
-    )[0][0]
+    return next((
+        v
+        for v, u in zip(data, data[1:])  # v = version, u = underline
+        if len(v) == len(u) and allSame(u) and hasDigit(v) and "." in v
+    ))